+++ Ranger/UI/TournamentTableWindow/TournamentViewController.swift
@@ -66,8 +66,7 @@ class TournamentViewController: NSViewController
         viewModel.update(with: tableWindowInfo)
         
         // UI.
-        if (App.configuration.isLiveMode)
-        { alignWindow(to: tableWindowInfo) }
+        alignWindow(to: tableWindowInfo)
     }
     
     func alignWindow(to tableWindowInfo: TableWindowInfo)